pyproject.toml
[project]
name = "dbtflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-duckdb>=1.9.6",
    "dbt-postgres>=1.9.1",
]
